improvement(map): pull out marker model into own package
